added improved GUI for user inputs, improved pot detection accuracy
