+++ unwarp_rectify/utils.py
@@ -9,6 +9,7 @@ from __future__ import absolute_import, division, print_function
 import numpy as np
 import cv2yml
 import cv2
+from matplotlib import pylab as plt
 
 def getRectCornersFrom2Points(Image, Points, AspectRatio, Rounded = False):
 #    print('Points =', Points)
@@ -164,3 +165,234 @@ def readCalibration(CalibFile):
     RVecs = parameters['RVecs']
     TVecs = parameters['TVecs']
     return ImageSize, SquareSize, CameraMatrix, DistCoefs, RVecs, TVecs
+
+def readGeometries(GeometryFile):
+    parameters = cv2yml.yml2dic(GeometryFile)
+    rotationAngle = parameters['rotationAngle']
+    distortionCorrected = bool(parameters['distortionCorrected'])
+    colorcardList = parameters['colorcardList'].tolist()
+    colorcardList2 = []
+    for i in range(0,len(colorcardList),4):
+        colorcardList2.append([colorcardList[i], colorcardList[i+1], \
+                               colorcardList[i+2], colorcardList[i+3]])
+    trayList = parameters['trayList'].tolist()
+    trayList2 = []
+    for i in range(0,len(trayList),4):
+        trayList2.append([trayList[i], trayList[i+1], \
+                          trayList[i+2], trayList[i+3]])
+    potList = parameters['potList'].tolist()
+    potList2 = []
+    for i in range(0,len(potList),4):
+        potList2.append([potList[i], potList[i+1], \
+                         potList[i+2], potList[i+3]])
+    return rotationAngle, distortionCorrected, colorcardList2, trayList2, potList2
+
+def createMap(Centre, Width, Height, Angle):
+    MapX, MapY = np.meshgrid(np.arange(Width), np.arange(Height))
+    MapX = MapX - Width/2.0
+    MapY = MapY - Height/2.0
+    MapX2 =  MapX*np.cos(Angle) + MapY*np.sin(Angle) + Centre[0]
+    MapY2 = -MapX*np.sin(Angle) + MapY*np.cos(Angle) + Centre[1]
+    return MapX2.astype(np.float32), MapY2.astype(np.float32)
+    
+# Using modified Gamma Correction Algorithm by
+# Constantinou2013 - A comparison of color correction algorithms for endoscopic cameras
+def getColorMatchingError(Arg, Colors, Captured_Colors):
+    ColorMatrix = Arg[:9].reshape([3,3])
+    ColorConstant = Arg[9:12]
+    ColorGamma = Arg[12:15]
+    ErrorList = []
+    for Color, Captured_Color in zip(Colors, Captured_Colors):
+        Color2 = np.dot(ColorMatrix, Captured_Color) + ColorConstant
+        Color3 = 255.0 * np.power(Color2/255.0, ColorGamma)
+        Error = np.linalg.norm(Color - Color3)
+        ErrorList.append(Error)
+    return ErrorList
+    
+def correctColor(Image, ColorMatrix, ColorConstant, ColorGamma):
+    ImageCorrected = np.zeros_like(Image)
+    for i in range(Image.shape[0]):
+        for j in range(Image.shape[1]):
+            Captured_Color = Image[i,j,:].reshape([3])
+            Color2 = np.dot(ColorMatrix, Captured_Color) + ColorConstant
+            Color3 = 255.0 * np.power(Color2/255.0, ColorGamma)
+            ImageCorrected[i,j,:] = np.uint8(Color3)
+    return ImageCorrected
+
+# Using modified Gamma Correction Algorithm by
+# Constantinou2013 - A comparison of color correction algorithms for endoscopic cameras
+def getColorMatchingErrorVectorised(Arg, Colors, Captured_Colors):
+    ColorMatrix = Arg[:9].reshape([3,3])
+    ColorConstant = Arg[9:12].reshape([3,1])
+    ColorGamma = Arg[12:15]
+    
+    TempRGB = np.dot(ColorMatrix, Captured_Colors) + ColorConstant
+    Corrected_Colors = np.zeros_like(TempRGB)
+    Corrected_Colors[0,:] = 255.0*np.power(TempRGB[0,:]/255.0, ColorGamma[0])
+    Corrected_Colors[1,:] = 255.0*np.power(TempRGB[1,:]/255.0, ColorGamma[1])
+    Corrected_Colors[2,:] = 255.0*np.power(TempRGB[2,:]/255.0, ColorGamma[2])
+    
+    Diff = Colors - Corrected_Colors
+    ErrorList = np.sqrt(np.sum(Diff*Diff, axis= 0)).tolist()
+    return ErrorList
+    
+def correctColorVectorised(Image, ColorMatrix, ColorConstant, ColorGamma):
+    Width, Height = Image.shape[1::-1]
+    CapturedR = Image[:,:,0].reshape([1,Width*Height])
+    CapturedG = Image[:,:,1].reshape([1,Width*Height])
+    CapturedB = Image[:,:,2].reshape([1,Width*Height])
+    CapturedRGB = np.concatenate((CapturedR, CapturedG, CapturedB), axis=0)
+    
+    TempRGB = np.dot(ColorMatrix, CapturedRGB) + ColorConstant
+    CorrectedRGB = np.zeros_like(TempRGB)
+    CorrectedRGB[0,:] = 255.0*np.power(TempRGB[0,:]/255.0, ColorGamma[0])
+    CorrectedRGB[1,:] = 255.0*np.power(TempRGB[1,:]/255.0, ColorGamma[1])
+    CorrectedRGB[2,:] = 255.0*np.power(TempRGB[2,:]/255.0, ColorGamma[2])
+    
+    CorrectedR = CorrectedRGB[0,:].reshape([Height, Width])
+    CorrectedG = CorrectedRGB[1,:].reshape([Height, Width])
+    CorrectedB = CorrectedRGB[2,:].reshape([Height, Width])
+    ImageCorrected = np.zeros_like(Image)
+    ImageCorrected[:,:,0] = CorrectedR
+    ImageCorrected[:,:,1] = CorrectedG
+    ImageCorrected[:,:,2] = CorrectedB
+    return ImageCorrected
+
+def rotateImage(Image, RotationAngle = 0.0):
+    Image_ = Image
+    if RotationAngle%90.0 == 0:
+        k = RotationAngle//90.0
+        Image_ = np.rot90(Image_, k)
+    elif RotationAngle != 0:
+        center=tuple(np.array(Image_.shape[0:2])/2)
+        rot_mat = cv2.getRotationMatrix2D(center, RotationAngle,1.0)
+        Image_ = cv2.warpAffine(Image_, rot_mat, Image_.shape[0:2],flags=cv2.INTER_LINEAR)
+    return Image_
+    
+def matchTemplateLocation(Image, Template, EstimatedLocation, SearchRange = [0.5, 0.5], RangeInImage = True):
+    if RangeInImage: # use image size
+        Width = Image.shape[1]
+        Height = Image.shape[0]
+    else: # use template size
+        Width = Template.shape[1]
+        Height = Template.shape[0]
+        
+    if SearchRange == None: # search throughout the whole images
+        CroppedHalfWidth  = Width//2 
+        CroppedHalfHeight = Height//2
+    elif SearchRange[0] <= 1.0 and SearchRange[1] <= 1.0: # in fraction values
+        CroppedHalfWidth  = (Template.shape[1]+SearchRange[0]*Width)//2 
+        CroppedHalfHeight = (Template.shape[0]+SearchRange[1]*Height)//2
+    else: # in pixels values
+        CroppedHalfWidth  = (Template.shape[1]+SearchRange[0])//2 
+        CroppedHalfHeight = (Template.shape[0]+SearchRange[1])//2
+        
+    if CroppedHalfWidth > Image.shape[1]//2-1:
+        CroppedHalfWidth = Image.shape[1]//2-1
+    if CroppedHalfHeight > Image.shape[0]//2-1:
+        CroppedHalfHeight = Image.shape[0]//2-1
+
+    SearchTopLeftCorner     = [EstimatedLocation[0]-CroppedHalfWidth, EstimatedLocation[1]-CroppedHalfHeight]
+    SearchBottomRightCorner = [EstimatedLocation[0]+CroppedHalfWidth, EstimatedLocation[1]+CroppedHalfHeight]
+
+    return matchTemplateWindow(Image, Template, SearchTopLeftCorner, SearchBottomRightCorner)
+
+def matchTemplateWindow(Image, Template, SearchTopLeftCorner, SearchBottomRightCorner):
+    CropedImage = Image[SearchTopLeftCorner[1]:SearchBottomRightCorner[1], SearchTopLeftCorner[0]:SearchBottomRightCorner[0]]
+    corrMap = cv2.matchTemplate(CropedImage.astype(np.uint8), Template.astype(np.uint8), cv2.TM_CCOEFF_NORMED)
+    _, maxVal, _, maxLoc = cv2.minMaxLoc(corrMap)
+    # recalculate max position in cropped image space
+    matchedLocImageCropped = (maxLoc[0] + Template.shape[1]//2, 
+                              maxLoc[1] + Template.shape[0]//2)
+    # recalculate max position in full image space
+    matchedLocImage = (matchedLocImageCropped[0] + SearchTopLeftCorner[0], \
+                       matchedLocImageCropped[1] + SearchTopLeftCorner[1])
+#    if isShow:
+#        plt.figure()
+#        plt.imshow(Template)
+#        plt.figure()
+#        plt.imshow(corrMap)
+#        plt.hold(True)
+#        plt.plot([maxLoc[0]], [maxLoc[1]], 'o')
+#        plt.figure()
+#        plt.imshow(CropedImage)
+#        plt.hold(True)
+#        plt.plot([matchedLocImageCropped[0]], [matchedLocImageCropped[1]], 'o')
+#        plt.figure()
+#        plt.imshow(Image)
+#        plt.hold(True)
+#        plt.plot([matchedLocImage[0]], [matchedLocImage[1]], 'o')
+#        plt.show()
+
+    return matchedLocImage, maxVal, maxLoc, corrMap
+    
+def createImagePyramid(Image, NoLevels = 5):
+    for i in range(NoLevels):
+        if i == 0:
+            PyramidImages = [Image.astype(np.uint8)]
+        else:
+            PyramidImages.append(cv2.pyrDown(PyramidImages[i-1]).astype(np.uint8))
+    return PyramidImages
+
+def matchTemplatePyramid(PyramidImages, PyramidTemplates, RotationAngle = None, \
+        EstimatedLocation = None, SearchRange = None, NoLevels = 4, FinalLevel = 1):
+    for i in range(NoLevels-1, -1, -1):
+        if i == NoLevels-1:
+            if EstimatedLocation == None:
+                maxLocEst = [PyramidImages[i].shape[1]//2, PyramidImages[i].shape[0]//2] # image center
+            else:
+                maxLocEst = [EstimatedLocation[0]//2**i, EstimatedLocation[1]//2**i] # scale position to the pyramid level
+
+            if SearchRange[0] > 1.0 and SearchRange[1] > 1.0:
+                SearchRange2 = [SearchRange[0]//2**i, SearchRange[1]//2**i]
+            else:
+                SearchRange2 = SearchRange
+
+            matchedLocImage, maxVal, maxLoc, corrMap = matchTemplateLocation(PyramidImages[i], PyramidTemplates[i], maxLocEst, SearchRange = SearchRange2)
+            if RotationAngle == None:
+                matchedLocImage180, maxVal180, maxLoc180, corrMap180 = matchTemplateLocation(np.rot90(PyramidImages[i],2).astype(np.uint8), PyramidTemplates[i], maxLocEst, SearchRange)
+                if maxVal < 0.3 and maxVal180 < 0.3:
+                    print('#### Warning: low matching score ####')
+#                    return None, None, None
+                if maxVal < maxVal180:
+                    PyramidImages = [np.rot90(Img,2) for Img in PyramidImages]
+                    matchedLocImage, matchedLocImage180 = matchedLocImage180, matchedLocImage
+                    maxVal, maxVal180 = maxVal180, maxVal
+                    maxLoc, maxLoc180 = maxLoc180, maxLoc
+                    corrMap, corrMap180 = corrMap180, corrMap
+                    RotationAngle = 180
+                else:
+                    RotationAngle = 0
+            # rescale to location in level-0 image
+            matchedLocImage0 = (matchedLocImage[0]*2**i, matchedLocImage[1]*2**i)
+        else:
+            maxLocEst = (matchedLocImage0[0]//2**i, matchedLocImage0[1]//2**i)
+            searchRange = [6,6]
+                            
+            matchedLocImage, maxVal, maxLoc, corrMap = matchTemplateLocation(PyramidImages[i], PyramidTemplates[i], maxLocEst, searchRange)
+            # rescale to location in level-0 image
+            matchedLocImage0 = (matchedLocImage[0]*2**i, matchedLocImage[1]*2**i)
+
+#        plt.figure()
+#        plt.imshow(PyramidTemplates[i])     
+#        
+#        plt.figure()
+#        plt.imshow(corrMap)
+#        plt.hold(True)
+#        plt.plot([maxLoc[0]], [maxLoc[1]], 'o')
+#        plt.title('maxVal = %f' %maxVal)
+#        
+#        plt.figure()
+#        plt.imshow(PyramidImages[i])
+#        plt.hold(True)
+#        plt.plot([matchedLocImage[0]], [matchedLocImage[1]], 'o')
+#        plt.plot([maxLocEst[0]], [maxLocEst[1]], 'x')
+#        plt.title('Level = %d, RotationAngle = %f' %(i, RotationAngle))
+#        plt.show()
+
+        if i ==  FinalLevel:
+            # Skip early to save time
+            break
+        
+    print('maxVal, maxLocImage, RotationAngle =', maxVal, matchedLocImage0, RotationAngle)
+    return maxVal, matchedLocImage0, RotationAngle
